+++ Sources/Server/Authentification/AuthCodeStorage+RedisImpl.swift
@@ -15,6 +15,7 @@ class RedisAuthCodeStorage: AuthCodeStorageProtocol {
         guard let key = RedisKey(rawValue: "\(session.type)~" + session.code.value) else {
             throw AuthCodeStorageError.KEY_ERROR
         }
+
         let sessionData = try JSONEncoder.main.encode(session)
         try redis.set(key, to: String(data: sessionData, encoding: .utf8)).wait()