+++ Sources/Server/ScriptingProvider/JavaScriptProvider.swift
@@ -25,6 +25,7 @@ class JavaScriptProvider: JSFunctionsDelegate {
     /// - See: committedResults
     ///
     private let group = DispatchGroup()
+    private let queue = DispatchQueue(label: "js_evaluate", qos: .default, attributes: .concurrent)
 
     /// The execution inside the javascript can not be monitored from outside the script itself. To get track of the
     /// state of the javascript, it has to `commit` its final results back into the caller stack once. A `DispatchGroup`
@@ -163,9 +164,10 @@ class JavaScriptProvider: JSFunctionsDelegate {
             class classToRun: ScriptClassExecution,
             arguments args: JSInputParameterProtocol?
     ) async throws -> [String?] {
-        try await withCheckedThrowingContinuation { continuation in
+        return try await withCheckedThrowingContinuation { continuation in
             start(class: classToRun, arguments: args, completion: continuation.resume(with:))
         }
+
     }
 
     /// Callback method to execute a specific function in the javascript context
@@ -184,7 +186,6 @@ class JavaScriptProvider: JSFunctionsDelegate {
             ) -> Void) {
         // enter the DispatchGroup to wait for a commit from within the javascript
         group.enter()
-
         do {
             // evaluate the script and get the result back. The result will be logged only, because the
             // script should run until the `commit` function is called.
@@ -200,14 +201,18 @@ class JavaScriptProvider: JSFunctionsDelegate {
                     Log.info("Provider script omitted result: \(stringValue)")
                 }
             }
+            if result.isBoolean {
+                let boolValue = result.booleanValue
+                Log.info("Provider script omitted boolean result: \(boolValue)")
+            }
         } catch {
             // toJSON is the only function that can throw an exception. No need to switch/case between error causes
             completion(.failure(.parserError(error.localizedDescription)))
             return
         }
-
         // wait that the DispatchGroup did notified and take the committed result
-        group.notify(queue: DispatchQueue.global(qos: .background)) { [self] in
+
+        group.notify(queue: queue) { [self] in
             if let committedResults = committedResults {
                 completion(.success(committedResults))
                 return
@@ -219,9 +224,11 @@ class JavaScriptProvider: JSFunctionsDelegate {
 
         // Wait until timeout for the script to evaluate till commit is called
         _ = group.wait(timeout: DispatchTime.now() + DispatchTimeInterval.seconds(Constants.PROVIDER.SCRIPT_TIMEOUT))
+
         if committedResults == nil {
-            completion(.failure(.timeout))
+             completion(.failure(.timeout))
         }
+        group.suspend()
     }
 
     /// Returns the property value from an initialized class as a Double