+++ Tests/ServerTests/ScriptingProvider/UserLoginProviderTests.swift
@@ -37,10 +37,22 @@ final class UserLoginProviderTests: XCTestCase {
     let userOK = JSInputCredentials(username: "ok@example.com", password: "very-secret")
     let userDenied = JSInputCredentials(username: "deni@example.com", password: "very-secret")
 
+    let app = Application(.testing)
+
+    override func setUp() {
+        super.setUp()
+        try? configure(app)
+    }
+
+    override func tearDown() {
+        app.shutdown()
+    }
+
     func testGetExampleCallback() async throws {
         let cbi = JavaScriptProvider()
         try cbi.loadProvider(script: providerScriptFetchExample)
         let expect = expectation(description: "userValidate response")
+
         cbi.start(class: .userLogin, arguments: userOK) { result in
             do {
                 let bodies = try result.get()
@@ -62,7 +74,7 @@ final class UserLoginProviderTests: XCTestCase {
                 }
             }
         }
-        wait(for: [expect], timeout: 10)
+        wait(for: [expect], timeout: TestDefaults.waitTimeout)
     }
 
     func testGetExampleAsync() async throws {