+++ Tests/ServerTests/setup-utils/TestDefaults.swift
@@ -1,0 +1,6 @@
+import XCTest
+@testable import Server
+
+public struct TestDefaults {
+    static let waitTimeout: TimeInterval = 90
+}